added buoyancy compensator code in IntegratedSteering.ino
The buoyancy compensator loop is integrated into the code. Pls check over the pin assignments and make sure they are plugged to the arduino correctly. The limit switch doesnt really matter which way.
Also motor technically doesnt matter as well, BUT pls do not start the compensator when the pin is close to the limit switch.